Refactor middleware constants (#9)
* feat: Add username validation middleware

This middleware and tests will return the 401 codes for missing JSON
entries. Assuming correct format, then it will validate the username
does NOT exist.

If it does then a 409 code is returned with a message explaining.

* feat: Add enum defining status codes for the API

* improve: Add individual middleware functions

These individual functions validate the JSON payload
but also allow reusable code for other endpoints.

* fix: replace magic numbers

The magic numbers are now replaced by the needed enum varient

* docs: prettier formatting

* docs: comments explaining verify funcs

* fix: add returns after sending resposne

* fix: remove variable on line 10

* fix: Remove debug and improve variable naming

* fix: improve readability

* chore: ran prettier

* fix: remove unneeded variable

* fix: put username var back

* fix: Add descriptive var names

* docs: Improve documentation and readability

In several locations I added the StatusCodes enum variant to replace
the local named constant OR magic number.

Function comment blocks were also added in a couple spots.

* refactor: create utils/

The status codes file is now located inside the utils/ directory in the API directory instead of just being present among everything. The sourcing was updated accordingly.

* refactor: user.ts login error message

The message now indicates that the credentials provided either are wrong or the account doesn't exist.